fixed issues with restructure merged
